refactor: 인증 화면과 앱 레이아웃 분리
인증 전용 레이아웃과 내부 서비스 전용 레이아웃을 분리해 라우트 책임을 명확히 하고 상위 App 구조를 단순하게 정리합니다.
